Replace the AsyncQueue.SemaphoreSlim with an AutoResetEvent. This addresses a deadlock that was happening at least on iOS, probably on Quest, and possibly very very rarely on desktop. I unfortunately don't know the actual cause of the deadlock, which doesn't make me fee great about this. The code all seemed fine. However the results are hard to ignore, it no longer eventually blocks on the semaphore.
This also slightly changes how the AsyncQueue works: The SemaphoreSlim meant that state (the queue count) was essentially in two places, the actual queue, and the semaphore. It was basically designed as push/pops always being paired. However in a traditional producer/consumer pattern I wouldn't expect either end to have any knowledge beyond what is in the queue, so this eliminates the stateful semaphore and just relies on the queue to tell us if there are any items in the queue.